requirements.txt
# Core ML / embeddings
numpy>=1.26.0
scikit-learn>=1.4.0
umap-learn>=0.5.6
hdbscan>=0.8.33

# Visualization
plotly>=5.20.0
matplotlib>=3.8.0
kaleido>=0.2.1          # static image export for plotly

# Data handling
pandas>=2.2.0
pydantic>=2.6.0

# HTTP (Ollama API)
httpx>=0.27.0

# CLI
click>=8.1.7
rich>=13.7.0             # pretty terminal output

# Config
pyyaml>=6.0.1

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.0
